requirements.txt
# Computer Vision
opencv-python-headless>=4.8.0
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Statistical Analysis
statsmodels>=0.14.0

# Utilities
pathlib2>=2.3.7
tqdm>=4.65.0
cupy-cuda115>=10.0.0,<11.0.0
